refactor: mutation 콜백 로직 훅으로 이동
