+++ src/components/molecules/input/Input.tsx
@@ -5,11 +5,11 @@ interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {
   rightIcon?: ReactNode;
 }
 
-export const Input = forwardRef<HTMLInputElement, InputProps>(({ leftIcon, rightIcon, ...props }, ref) => {
+export const Input = forwardRef<HTMLInputElement, InputProps>(({ name, leftIcon, rightIcon, ...props }, ref) => {
   return (
     <div className='flex w-full h-[50px] rounded-[10px] p-[8px] border border-gray-300 align-middle items-center'>
       {leftIcon && leftIcon}
-      <input ref={ref} className='w-full p-[4px]' {...props} />
+      <input name={name} ref={ref} className='w-full p-[4px]' {...props} />
       {rightIcon && rightIcon}
     </div>
   );
